removed a couple regression issues by restoring older code versions.

+++ public/js/pickplayers.js
@@ -1,37 +1,33 @@
 
 
 var arrayPlayerOfIds = [];
+
+
 var deleteGuy = function(gameid, guy){
 
-		$.getJSON("/api/gamePlayer/" + gameid, function( data ) {
-			arrayPlayerOfIds = data;
-			console.log(guy, arrayPlayerOfIds.toString());
-		});
-		for (var i = 0; i < arrayPlayerOfIds; i++) {
-			console.log("inside for loop")
-			console.log(arrayPlayerOfIds[i].toString())
-			if(guy === arrayPlayerOfIds[i].toString()){
-			console.log("preparing to enter ajax", gameid, guy)
+	$.getJSON("/api/gamePlayer/" + gameid, function( data ) {
+		var foundGuy = null;
+		for (var i = 0; i < data.length; i++) {	
+		  	if( data[i] === guy ) {
+				foundGuy = guy;
+			};
+		};
+		if (foundGuy != null) {
 			$.ajax({
-				url: "/api/gameRoutes/" + gameid,
+				url: "/api/gamePlayer/" + gameid,
 				type: 'DELETE',
 				dataType: 'json',
-				data: { "_players" : 'ObjectId("' + arrayPlayerOfIds[i] + '")'},
-
+				data: { "player": guy},
 				success: function(result) {
-
-					console.log(result)
         			console.log('Successfully removed ' + guy);	
-        			location.href = "/game/" + gameid;
-    			},
-    			error: function(err){
-    				console.log(err)
+        			location.href = "/completeGame/" + gameid;
     			}
 			});
-		}
-	};
+		} else {
+			console.log("Couldn't find " + guy);
+		};	
+	});
 };
-
 
 
 var finalizeGame = function (){
